requirements.txt
openai
langchain
langgraph
langchain-openai
langchain_community
langchain-chroma
chromadb
tiktoken
beautifulsoup4 
html2text 
pdfplumber 
pypdf 
python-docx
nltk 
sentence-transformers
tqdm 
loguru 
python-dotenv 
pydantic 
httpx
black

